refactor(@angular-devkit/build-angular): move `SourceFileCache` into dedicated file
This is to reduce the code in `compiler-plugin.ts`
