+++ packages/angular_devkit/build_angular/src/tools/esbuild/angular/compiler-plugin.ts
@@ -16,13 +16,12 @@ import type {
 } from 'esbuild';
 import assert from 'node:assert';
 import { realpath } from 'node:fs/promises';
-import { platform } from 'node:os';
 import * as path from 'node:path';
 import { pathToFileURL } from 'node:url';
 import ts from 'typescript';
 import { maxWorkers } from '../../../utils/environment-options';
 import { JavaScriptTransformer } from '../javascript-transformer';
-import { LoadResultCache, MemoryLoadResultCache } from '../load-result-cache';
+import { LoadResultCache } from '../load-result-cache';
 import {
   logCumulativeDurations,
   profileAsync,
@@ -33,39 +32,7 @@ import { BundleStylesheetOptions, bundleComponentStylesheet } from '../styleshee
 import { AngularHostOptions } from './angular-host';
 import { AngularCompilation, AotCompilation, JitCompilation, NoopCompilation } from './compilation';
 import { setupJitPluginCallbacks } from './jit-plugin-callbacks';
-
-const USING_WINDOWS = platform() === 'win32';
-const WINDOWS_SEP_REGEXP = new RegExp(`\\${path.win32.sep}`, 'g');
-
-export class SourceFileCache extends Map<string, ts.SourceFile> {
-  readonly modifiedFiles = new Set<string>();
-  readonly babelFileCache = new Map<string, Uint8Array>();
-  readonly typeScriptFileCache = new Map<string, string | Uint8Array>();
-  readonly loadResultCache = new MemoryLoadResultCache();
-
-  referencedFiles?: readonly string[];
-
-  constructor(readonly persistentCachePath?: string) {
-    super();
-  }
-
-  invalidate(files: Iterable<string>): void {
-    this.modifiedFiles.clear();
-    for (let file of files) {
-      this.babelFileCache.delete(file);
-      this.typeScriptFileCache.delete(pathToFileURL(file).href);
-      this.loadResultCache.invalidate(file);
-
-      // Normalize separators to allow matching TypeScript Host paths
-      if (USING_WINDOWS) {
-        file = file.replace(WINDOWS_SEP_REGEXP, path.posix.sep);
-      }
-
-      this.delete(file);
-      this.modifiedFiles.add(file);
-    }
-  }
-}
+import { SourceFileCache } from './source-file-cache';
 
 export interface CompilerPluginOptions {
   sourcemap: boolean;
