+++ packages/angular_devkit/build_angular/src/tools/esbuild/angular/source-file-cache.ts
@@ -1,0 +1,46 @@
+/**
+ * @license
+ * Copyright Google LLC All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+
+import { platform } from 'node:os';
+import * as path from 'node:path';
+import { pathToFileURL } from 'node:url';
+import ts from 'typescript';
+import { MemoryLoadResultCache } from '../load-result-cache';
+
+const USING_WINDOWS = platform() === 'win32';
+const WINDOWS_SEP_REGEXP = new RegExp(`\\${path.win32.sep}`, 'g');
+
+export class SourceFileCache extends Map<string, ts.SourceFile> {
+  readonly modifiedFiles = new Set<string>();
+  readonly babelFileCache = new Map<string, Uint8Array>();
+  readonly typeScriptFileCache = new Map<string, string | Uint8Array>();
+  readonly loadResultCache = new MemoryLoadResultCache();
+
+  referencedFiles?: readonly string[];
+
+  constructor(readonly persistentCachePath?: string) {
+    super();
+  }
+
+  invalidate(files: Iterable<string>): void {
+    this.modifiedFiles.clear();
+    for (let file of files) {
+      this.babelFileCache.delete(file);
+      this.typeScriptFileCache.delete(pathToFileURL(file).href);
+      this.loadResultCache.invalidate(file);
+
+      // Normalize separators to allow matching TypeScript Host paths
+      if (USING_WINDOWS) {
+        file = file.replace(WINDOWS_SEP_REGEXP, path.posix.sep);
+      }
+
+      this.delete(file);
+      this.modifiedFiles.add(file);
+    }
+  }
+}
