+++ packages/angular_devkit/build_angular/src/tools/esbuild/application-code-bundle.ts
@@ -13,7 +13,8 @@ import { readFile } from 'node:fs/promises';
 import { extname, join, relative } from 'node:path';
 import type { NormalizedApplicationBuildOptions } from '../../builders/application/options';
 import { allowMangle } from '../../utils/environment-options';
-import { SourceFileCache, createCompilerPlugin } from './angular/compiler-plugin';
+import { createCompilerPlugin } from './angular/compiler-plugin';
+import { SourceFileCache } from './angular/source-file-cache';
 import { createCompilerPluginOptions } from './compiler-plugin-options';
 import { createAngularLocaleDataPlugin } from './i18n-locale-plugin';
 import { createRxjsEsmResolutionPlugin } from './rxjs-esm-resolution-plugin';
