+++ packages/angular_devkit/build_angular/src/tools/esbuild/bundler-execution-result.ts
@@ -7,7 +7,7 @@
  */
 
 import type { ChangedFiles } from '../../tools/esbuild/watcher';
-import type { SourceFileCache } from './angular/compiler-plugin';
+import type { SourceFileCache } from './angular/source-file-cache';
 import type { BuildOutputFile, BuildOutputFileType, BundlerContext } from './bundler-context';
 import { createOutputFileFromText } from './utils';
