+++ packages/angular_devkit/build_angular/src/tools/esbuild/compiler-plugin-options.ts
@@ -7,7 +7,8 @@
  */
 
 import { NormalizedApplicationBuildOptions } from '../../builders/application/options';
-import type { SourceFileCache, createCompilerPlugin } from './angular/compiler-plugin';
+import type { createCompilerPlugin } from './angular/compiler-plugin';
+import type { SourceFileCache } from './angular/source-file-cache';
 
 type CreateCompilerPluginParameters = Parameters<typeof createCompilerPlugin>;
